+++ include/stdexec/functional.hpp
@@ -17,6 +17,8 @@
 
 #include "__detail/__config.hpp"
 #include "__detail/__meta.hpp"
+#include "__detail/__tag_invoke.hpp"
+
 #include "concepts.hpp"
 
 #include <functional>
@@ -35,10 +37,6 @@ namespace stdexec::__std_concepts {
     };
 #endif
 } // namespace stdexec::__std_concepts
-
-namespace std {
-  using namespace stdexec::__std_concepts;
-} // namespace std
 
 namespace stdexec {
   template <auto _Fun>
@@ -326,75 +324,4 @@ namespace stdexec {
 
   template <class _Tag>
   inline constexpr __mkfield_<_Tag> __mkfield{};
-
-  // [func.tag_invoke], tag_invoke
-  namespace __tag_invoke {
-    void tag_invoke();
-
-    // NOT TO SPEC: Don't require tag_invocable to subsume invocable.
-    // std::invoke is more expensive at compile time than necessary,
-    // and results in diagnostics that are more verbose than necessary.
-    template <class _Tag, class... _Args>
-    concept tag_invocable = //
-      requires(_Tag __tag, _Args&&... __args) {
-        tag_invoke(static_cast<_Tag&&>(__tag), static_cast<_Args&&>(__args)...);
-      };
-
-    template <class _Ret, class _Tag, class... _Args>
-    concept __tag_invocable_r = //
-      requires(_Tag __tag, _Args&&... __args) {
-        {
-          static_cast<_Ret>(tag_invoke(static_cast<_Tag&&>(__tag), static_cast<_Args&&>(__args)...))
-        };
-      };
-
-    // NOT TO SPEC: nothrow_tag_invocable subsumes tag_invocable
-    template <class _Tag, class... _Args>
-    concept nothrow_tag_invocable =
-      tag_invocable<_Tag, _Args...> && //
-      requires(_Tag __tag, _Args&&... __args) {
-        { tag_invoke(static_cast<_Tag&&>(__tag), static_cast<_Args&&>(__args)...) } noexcept;
-      };
-
-    template <class _Tag, class... _Args>
-    using tag_invoke_result_t = decltype(tag_invoke(__declval<_Tag>(), __declval<_Args>()...));
-
-    template <class _Tag, class... _Args>
-    struct tag_invoke_result { };
-
-    template <class _Tag, class... _Args>
-      requires tag_invocable<_Tag, _Args...>
-    struct tag_invoke_result<_Tag, _Args...> {
-      using type = tag_invoke_result_t<_Tag, _Args...>;
-    };
-
-    struct tag_invoke_t {
-      template <class _Tag, class... _Args>
-        requires tag_invocable<_Tag, _Args...>
-      STDEXEC_ATTRIBUTE((always_inline))
-      constexpr auto
-        operator()(_Tag __tag, _Args&&... __args) const
-        noexcept(nothrow_tag_invocable<_Tag, _Args...>) -> tag_invoke_result_t<_Tag, _Args...> {
-        return tag_invoke(static_cast<_Tag&&>(__tag), static_cast<_Args&&>(__args)...);
-      }
-    };
-
-  } // namespace __tag_invoke
-
-  using __tag_invoke::tag_invoke_t;
-
-  namespace __ti {
-    inline constexpr tag_invoke_t tag_invoke{};
-  } // namespace __ti
-
-  using namespace __ti;
-
-  template <auto& _Tag>
-  using tag_t = __decay_t<decltype(_Tag)>;
-
-  using __tag_invoke::tag_invocable;
-  using __tag_invoke::__tag_invocable_r;
-  using __tag_invoke::nothrow_tag_invocable;
-  using __tag_invoke::tag_invoke_result_t;
-  using __tag_invoke::tag_invoke_result;
 } // namespace stdexec
